11 simuation service implementation (#12)
#11 simulation service

+++ rac-commons/src/main/java/pl/mg/rac/commons/value/Location.java
@@ -1,7 +1,9 @@
 package pl.mg.rac.commons.value;
 
+import java.math.BigDecimal;
+
 public record Location(
-        double latitude,
-        double longitude
+        BigDecimal latitude,
+        BigDecimal longitude
 ) {
 }
